src/daft/requirements.txt
# Core dependencies for OCR pipeline (Gemini)
google-generativeai>=0.7.2
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
tqdm>=4.66.0
requests>=2.31.0

# Image/PDF handling
opencv-python>=4.9.0.80
Pillow>=10.3.0
pdf2image>=1.17.0
scikit-image>=0.23.2
scipy>=1.12.0

# Optional: lightweight API/CLI enhancements
typer>=0.12.3
rich>=13.7.1

# Word document export
python-docx>=1.1.0


torchvision
matplotlib
addict
torch==2.6.0
transformers==4.46.3
tokenizers==0.20.3
einops
easydict

# DeepSeek-VL2 for OCR and object detection
deepseek-vl2